@@ -3,70 +3,78 @@
   "version": "0.1.0",
   "private": true,
   "scripts": {
-    "dev": "yarn tinacms server:start -c \"next dev\"",
-    "build": "yarn tinacms server:start -c \"next build\"",
-    "start": "yarn tinacms server:start -c \"next start\"",
+    "dev": "cross-env NODE_ENV=development yarn tinacms dev --datalayer-port 9001 -c \"next dev\"",
+    "build": "cross-env NODE_ENV=production yarn tinacms dev --datalayer-port 9001 -c \"next build\"",
+    "start": "next start",
     "lint": "next lint"
   },
   "dependencies": {
     "@accessible/drawer": "^3.0.2",
+    "@babel/core": "^7.24.6",
+    "@babel/plugin-syntax-flow": "^7.24.6",
+    "@babel/plugin-transform-react-jsx": "^7.24.6",
     "@n8tb1t/use-scroll-position": "^2.0.3",
     "@sendgrid/mail": "^7.7.0",
-    "@svgr/webpack": "^8.0.0",
+    "@svgr/webpack": "^8.1.0",
+    "cross-env": "^7.0.3",
     "css-in-js-media": "^2.0.1",
-    "date-fns": "^2.29.3",
+    "date-fns": "^2.30.0",
     "gray-matter": "^4.0.3",
     "lodash": "^4.17.21",
-    "next": "12.1.0",
+    "next": "14.2.3",
     "nextjs-color-mode": "^1.0.5",
-    "polished": "^4.1.3",
+    "polished": "^4.3.1",
     "prism-react-renderer": "^1.3.5",
-    "react": "17.0.2",
-    "react-animate-height": "^2.0.23",
-    "react-dom": "17.0.2",
-    "react-hook-form": "^7.17.4",
+    "react": "18.2.0",
+    "react-animate-height": "^2.1.2",
+    "react-dom": "18.2.0",
+    "react-hook-form": "^7.51.5",
+    "react-is": "18.2.0",
     "react-mailchimp-subscribe": "^2.1.3",
     "react-schemaorg": "^2.0.0",
     "react-share": "^4.4.1",
     "reading-time": "^1.5.0",
     "schema-dts": "^1.1.2",
-    "styled-components": "^5.3.5",
-    "swiper": "9.3.2",
-    "tinacms": "^1.5.8",
+    "styled-components": "^5.3.11",
+    "swiper": "9.4.1",
+    "tinacms": "^1.6.3",
     "url-loader": "^4.1.1",
     "use-clipboard-copy": "^0.2.0",
-    "use-resize-observer": "^9.1.0"
+    "use-resize-observer": "^9.1.0",
+    "webpack": "^5.91.0"
   },
   "devDependencies": {
-    "@babel/eslint-parser": "^7.11.0",
+    "@babel/eslint-parser": "^7.24.6",
     "@fec/remark-a11y-emoji": "^3.1.0",
-    "@next/bundle-analyzer": "^13.3.1",
-    "@tinacms/cli": "^0.60.0",
-    "@types/react": "^17.0.20",
-    "@types/react-mailchimp-subscribe": "^2.1.1",
-    "@types/styled-components": "^5.1.25",
+    "@next/bundle-analyzer": "^13.5.6",
+    "@tinacms/cli": "^1.5.45",
+    "@types/babel__core": "^7.20.5",
+    "@types/react": "18.3.3",
+    "@types/react-is": "18.3.0",
+    "@types/react-mailchimp-subscribe": "^2.1.4",
+    "@types/styled-components": "^5.1.34",
     "@types/swiper": "^6.0.0",
-    "@typescript-eslint/eslint-plugin": "^5.0.0",
-    "@typescript-eslint/parser": "^5.59.9",
-    "all-contributors-cli": "^6.26.0",
+    "@typescript-eslint/eslint-plugin": "^5.62.0",
+    "@typescript-eslint/parser": "^5.62.0",
+    "all-contributors-cli": "^6.26.1",
     "copy-webpack-plugin": "^11.0.0",
-    "eslint": "^8.0.0",
-    "eslint-config-next": "^13.2.4",
-    "eslint-config-prettier": "^8.3.0",
-    "eslint-config-react-app": "^7.0.0",
-    "eslint-plugin-flowtype": "^8.0.0",
-    "eslint-plugin-import": "^2.24.2",
-    "eslint-plugin-jsx-a11y": "^6.7.1",
-    "eslint-plugin-react": "^7.25.1",
-    "eslint-plugin-react-hooks": "^4.6.0",
+    "eslint": "^8.57.0",
+    "eslint-config-next": "^13.5.6",
+    "eslint-config-prettier": "^8.10.0",
+    "eslint-config-react-app": "^7.0.1",
+    "eslint-plugin-flowtype": "^8.0.3",
+    "eslint-plugin-import": "^2.29.1",
+    "eslint-plugin-jsx-a11y": "^6.8.0",
+    "eslint-plugin-react": "^7.34.1",
+    "eslint-plugin-react-hooks": "^4.6.2",
     "next-mdx-remote": "^4.4.1",
-    "remark-breaks": "^3.0.1",
+    "remark-breaks": "^3.0.3",
     "remark-external-links": "^9.0.1",
     "remark-footnotes": "^4.0.1",
     "remark-gfm": "^3.0.1",
     "remark-sectionize": "^2.0.0",
     "remark-slug": "^7.0.1",
-    "typescript": "5.1.6"
+    "typescript": "5.4.5"
   },
-  "packageManager": "yarn@4.2.2"
+  "packageManager": "yarn@1.22.22"
 }
