@@ -8,7 +8,7 @@ export interface LinkProps {
 
 export default function Link({ href, children }: PropsWithChildren<LinkProps>) {
   return (
-    <NextLink href={href} passHref>
+    <NextLink href={href} passHref legacyBehavior>
       <Anchor>{children}</Anchor>
     </NextLink>
   );
