@@ -1,4 +1,4 @@
-import NextImage from 'next/image';
+import NextImage from "next/image";
 import React from 'react';
 import styled from 'styled-components';
 import { Autoplay } from 'swiper';
@@ -36,7 +36,15 @@ export default function Partners() {
       >
         {PARTNER_LOGOS.map((logo) => (
           <SwiperSlide key={logo}>
-            <NextImage src={'/partners/' + logo} alt={normalizePartnerLogoName(logo)} width={128} height={128} />
+            <NextImage
+              src={'/partners/' + logo}
+              alt={normalizePartnerLogoName(logo)}
+              width={128}
+              height={128}
+              style={{
+                maxWidth: "100%",
+                height: "auto"
+              }} />
           </SwiperSlide>
         ))}
       </Swiper>
